# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation for SMIL",
        "contact": {
            "email": "muhammad.raflysadewa.tik17@mhsw.pnj.ac.id"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/detail-alat": {
            "get": {
                "tags": [
                    "Detail Alat"
                ],
                "summary": "Return List of Detail Alat",
                "operationId": "getAllDetailAlat",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Detail Alat"
                ],
                "summary": "Create New Detail Alat",
                "operationId": "createNewDetailAlat",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreDetailAlatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetailAlatDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/filter/detail-alat/{alatId}": {
            "post": {
                "tags": [
                    "Detail Alat"
                ],
                "summary": "Return List of Detail Alat by Filter",
                "operationId": "getAllDetailAlatByFilter",
                "parameters": [
                    {
                        "name": "alatId",
                        "in": "path",
                        "description": "Alat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FilterDetailAlatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetailAlatListResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/detail-alat/get-by-alat-id/{alatId}": {
            "get": {
                "tags": [
                    "Detail Alat"
                ],
                "summary": "Get Detail Alat by Alat Id Information",
                "operationId": "getAllDetailAlatByAlatId",
                "parameters": [
                    {
                        "name": "alatId",
                        "in": "path",
                        "description": "Detail Alat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetailAlatDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/detail-alat/{id}": {
            "get": {
                "tags": [
                    "Detail Alat"
                ],
                "summary": "Get Detail Alat Detail Information",
                "operationId": "getAllDetailAlatById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Detail Alat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetailAlatDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Detail Alat"
                ],
                "summary": "Delete Existed DetailAlat",
                "operationId": "deleteExistingDetailAlat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "DetailAlat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/detail-alat/update-condition/{id}": {
            "put": {
                "tags": [
                    "Detail Alat"
                ],
                "summary": "Update  Condition Detail Alat ",
                "operationId": "update Condition Detail Alat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Detail Alat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "condition_status"
                                ],
                                "properties": {
                                    "condition_status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetailAlatDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/detail-alat/update-available/{id}": {
            "put": {
                "tags": [
                    "Detail Alat"
                ],
                "summary": "Update  Available Status Detail Alat ",
                "operationId": "update Available Status Detail Alat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Detail Alat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "available_status"
                                ],
                                "properties": {
                                    "available_status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetailAlatDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreDetailAlatRequest": {
                "title": "Store DetailAlat Request",
                "description": "Store DetailAlat Request body data",
                "required": [
                    "jabatan_name"
                ],
                "properties": {
                    "alat_id": {
                        "title": "alat_id",
                        "description": "Alat ID",
                        "type": "integer"
                    },
                    "lokasi_id": {
                        "title": "lokasi_id",
                        "description": "Lokasi ID",
                        "type": "integer"
                    },
                    "total_alat": {
                        "title": "total_alat",
                        "description": "Jumlah Alat",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "DetailAlatDetailResource": {
                "title": "DetailAlatDetailResource",
                "description": "Prodi Resource",
                "properties": {
                    "response": {
                        "$ref": "#/components/schemas/Response"
                    },
                    "data": {
                        "title": "Data",
                        "description": "Data Prodi Wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DetailAlat"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DetailAlatDetailResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}